+++ src/idunno.Bluesky/Actor/Agent/Helpers/GetStatus.cs
@@ -1,0 +1,32 @@
+// Copyright (c) Barry Dorrans. All rights reserved.
+// Licensed under the MIT License.
+
+using idunno.AtProto;
+using idunno.AtProto.Repo;
+using idunno.Bluesky.Actor;
+
+namespace idunno.Bluesky;
+
+public partial class BlueskyAgent
+{
+    /// <summary>
+    /// Gets the current profile status, if any, for the authenticated user.
+    /// </summary>
+    /// <param name="cancellationToken">A cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
+    /// <returns>The task object representing the asynchronous operation.</returns>
+    /// <exception cref="AuthenticationRequiredException">Thrown when the current agent is not authenticated.</exception>
+    public async Task<AtProtoHttpResult<AtProtoRepositoryRecord<Status>>> GetStatus(CancellationToken cancellationToken = default)
+    {
+        if (!IsAuthenticated)
+        {
+            throw new AuthenticationRequiredException();
+        }
+
+        AtUri statusUri = new($"at://{Did}/{CollectionNsid.Status}/self");
+        AtProtoHttpResult<AtProtoRepositoryRecord<Status>> getResult = await GetBlueskyRecord<Status>(
+            uri: statusUri,
+            cancellationToken: cancellationToken).ConfigureAwait(false);
+
+        return getResult;
+    }
+}
