+++ src/idunno.Bluesky/Record/BlueskyAgent.cs
@@ -3,10 +3,10 @@
 
 using System.Diagnostics.CodeAnalysis;
 using System.Net;
-using System.Text.Json;
 
 using idunno.AtProto;
 using idunno.AtProto.Repo;
+using idunno.Bluesky.Actor;
 using idunno.Bluesky.Embed;
 using idunno.Bluesky.Graph;
 using idunno.Bluesky.Record;
@@ -222,36 +222,6 @@ public partial class BlueskyAgent
     }
 
     /// <summary>
-    /// Gets a <see cref="Profile"/> for the current user.
-    /// </summary>
-    /// <param name="cancellationToken">A cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
-    /// <returns>The task object representing the asynchronous operation.</returns>
-    /// <exception cref="AuthenticationRequiredException">Thrown when the current agent is not authenticated.</exception>
-    [UnconditionalSuppressMessage(
-        "Trimming",
-        "IL2026:Members annotated with 'RequiresUnreferencedCodeAttribute' require dynamic access otherwise can break functionality when trimming application code",
-        Justification = "All types are preserved in the JsonSerializerOptions call to Get().")]
-    [UnconditionalSuppressMessage("AOT",
-        "IL3050:Calling members annotated with 'RequiresDynamicCodeAttribute' may break functionality when AOT compiling.",
-        Justification = "All types are preserved in the JsonSerializerOptions call to Get().")]
-    public async Task<AtProtoHttpResult<AtProtoRepositoryRecord<Profile>>> GetProfileRecord(
-        CancellationToken cancellationToken = default)
-    {
-        if (!IsAuthenticated)
-        {
-            throw new AuthenticationRequiredException();
-        }
-
-        AtUri profileUri = new($"at://{Did}/{CollectionNsid.Profile}/self");
-
-        return await GetRecord<Profile>(
-            profileUri,
-            service: Service,
-            jsonSerializerOptions: BlueskyServer.BlueskyJsonSerializerOptions,
-            cancellationToken: cancellationToken).ConfigureAwait(false);
-    }
-
-    /// <summary>
     /// Gets a page of records in the specified <paramref name="collection"/>.
     /// </summary>
     /// <typeparam name="TRecord">The type of the record value to get.</typeparam>
@@ -295,169 +265,6 @@ public partial class BlueskyAgent
             cancellationToken: cancellationToken).ConfigureAwait(false);
     }
 
-
-    /// <summary>
-    /// Sets the current user's <see cref="Profile"/>.
-    /// </summary>
-    /// <param name="profile">The <see cref="Profile"/> to create.</param>
-    /// <param name="cancellationToken">A cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
-    /// <returns>The task object representing the asynchronous operation.</returns>
-    /// <exception cref="ArgumentNullException">Thrown when <paramref name="profile"/> is <see langword="null"/>.</exception>
-    /// <exception cref="AuthenticationRequiredException">Thrown when the current agent is not authenticated.</exception>
-    [UnconditionalSuppressMessage(
-    "Trimming",
-    "IL2026:Members annotated with 'RequiresUnreferencedCodeAttribute' require dynamic access otherwise can break functionality when trimming application code",
-    Justification = "All types are preserved in the JsonSerializerOptions call to Put().")]
-    [UnconditionalSuppressMessage("AOT",
-    "IL3050:Calling members annotated with 'RequiresDynamicCodeAttribute' may break functionality when AOT compiling.",
-    Justification = "All types are preserved in the JsonSerializerOptions call to Put().")]
-    public async Task<AtProtoHttpResult<CreateRecordResult>> CreateProfile(
-    Profile profile,
-    CancellationToken cancellationToken = default)
-    {
-        ArgumentNullException.ThrowIfNull(profile);
-
-        if (!IsAuthenticated)
-        {
-            throw new AuthenticationRequiredException();
-        }
-
-        return await CreateBlueskyRecord(
-            record: profile,
-            collection: CollectionNsid.Profile,
-            rKey: "self",
-            cancellationToken: cancellationToken).ConfigureAwait(false);
-    }
-
-    /// <summary>
-    /// Creates or updates the current users profile.
-    /// </summary>
-    /// <param name="profile">The profile to create from or update to</param>
-    /// <param name="cancellationToken">A cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
-    /// <returns>The task object representing the asynchronous operation.</returns>
-    /// <exception cref="ArgumentNullException">Thrown when <paramref name="profile"/> is <see langword="null"/>.</exception>
-    /// <exception cref="AuthenticationRequiredException">Thrown when the current agent is not authenticated.</exception>
-    public async Task<AtProtoHttpResult<PutRecordResult>> SetProfile(
-        Profile profile,
-        CancellationToken cancellationToken = default)
-    {
-        ArgumentNullException.ThrowIfNull(profile);
-
-        if (!IsAuthenticated)
-        {
-            throw new AuthenticationRequiredException();
-        }
-
-        return await UpdateProfile(profile, cancellationToken).ConfigureAwait(false);
-    }
-
-    /// <summary>
-    /// Updates the current users profile.
-    /// </summary>
-    /// <param name="profile">The profile update to</param>
-    /// <param name="cancellationToken">A cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
-    /// <returns>The task object representing the asynchronous operation.</returns>
-    /// <exception cref="ArgumentNullException">Thrown when <paramref name="profile"/>, or its Value is <see langword="null"/>.</exception>
-    /// <exception cref="AuthenticationRequiredException">Thrown when the current agent is not authenticated.</exception>
-    public async Task<AtProtoHttpResult<PutRecordResult>> SetProfile(
-        AtProtoRepositoryRecord<Profile> profile,
-        CancellationToken cancellationToken = default)
-    {
-        ArgumentNullException.ThrowIfNull(profile);
-        ArgumentNullException.ThrowIfNull(profile.Value);
-
-        if (!IsAuthenticated)
-        {
-            throw new AuthenticationRequiredException();
-        }
-
-        return await UpdateProfile(profile, cancellationToken).ConfigureAwait(false);
-    }
-
-    /// <summary>
-    /// Update the current user's <see cref="Profile"/>.
-    /// </summary>
-    /// <param name="profile">The <see cref="AtProtoRepositoryRecord{Profile}"/> to update.</param>
-    /// <param name="cancellationToken">A cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
-    /// <returns>The task object representing the asynchronous operation.</returns>
-    /// <exception cref="ArgumentException">Thrown when the <paramref name="profile"/> is not valid.</exception>
-    /// <exception cref="AuthenticationRequiredException">Thrown when the current agent is not authenticated.</exception>
-    [UnconditionalSuppressMessage(
-        "Trimming",
-        "IL2026:Members annotated with 'RequiresUnreferencedCodeAttribute' require dynamic access otherwise can break functionality when trimming application code",
-        Justification = "All types are preserved in the JsonSerializerOptions call to Put().")]
-    [UnconditionalSuppressMessage("AOT",
-        "IL3050:Calling members annotated with 'RequiresDynamicCodeAttribute' may break functionality when AOT compiling.",
-        Justification = "All types are preserved in the JsonSerializerOptions call to Put().")]
-    public async Task<AtProtoHttpResult<PutRecordResult>> UpdateProfile(
-        AtProtoRepositoryRecord<Profile> profile,
-        CancellationToken cancellationToken = default)
-    {
-        ArgumentNullException.ThrowIfNull(profile);
-        ArgumentNullException.ThrowIfNull(profile.Value);
-
-        if (!IsAuthenticated)
-        {
-            throw new AuthenticationRequiredException();
-        }
-
-        if (profile.Uri.Authority is not Did recordDid)
-        {
-            throw new ArgumentException("Uri authority is not a DID", nameof(profile));
-        }
-
-        if (recordDid != Did)
-        {
-            throw new ArgumentException("Uri authority does not match the current user", nameof(profile));
-        }
-
-        return await PutRecord(
-            record: profile.Value,
-            jsonSerializerOptions: BlueskyServer.BlueskyJsonSerializerOptions,
-            collection: CollectionNsid.Profile,
-            rKey: "self",
-            validate: null,
-            swapCommit: null,
-            swapRecord: profile.Cid,
-            cancellationToken: cancellationToken).ConfigureAwait(false);
-    }
-
-    /// <summary>
-    /// Update the current user's <see cref="Profile"/>.
-    /// </summary>
-    /// <param name="profile">The <see cref="Profile"/> to update with.</param>
-    /// <param name="cancellationToken">A cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
-    /// <returns>The task object representing the asynchronous operation.</returns>
-    /// <exception cref="ArgumentNullException">Thrown when <paramref name="profile"/> is <see langword="null"/>.</exception>
-    /// <exception cref="AuthenticationRequiredException">Thrown when the current agent is not authenticated.</exception>
-    [UnconditionalSuppressMessage(
-        "Trimming",
-        "IL2026:Members annotated with 'RequiresUnreferencedCodeAttribute' require dynamic access otherwise can break functionality when trimming application code",
-        Justification = "All types are preserved in the JsonSerializerOptions call to Put().")]
-    [UnconditionalSuppressMessage("AOT",
-        "IL3050:Calling members annotated with 'RequiresDynamicCodeAttribute' may break functionality when AOT compiling.",
-        Justification = "All types are preserved in the JsonSerializerOptions call to Put().")]
-    public async Task<AtProtoHttpResult<PutRecordResult>> UpdateProfile(
-        Profile profile,
-        CancellationToken cancellationToken = default)
-    {
-        ArgumentNullException.ThrowIfNull(profile);
-
-        if (!IsAuthenticated)
-        {
-            throw new AuthenticationRequiredException();
-        }
-
-        return await PutRecord(
-            record: profile,
-            jsonSerializerOptions: BlueskyServer.BlueskyJsonSerializerOptions,
-            collection: CollectionNsid.Profile,
-            rKey: "self",
-            validate: null,
-            swapCommit: null,
-            swapRecord: null,
-            cancellationToken: cancellationToken).ConfigureAwait(false);
-    }
 
     /// <summary>
     /// Creates a <see cref="List"/>.
@@ -630,7 +437,6 @@ public partial class BlueskyAgent
         ArgumentNullException.ThrowIfNull(uri.Collection);
         ArgumentOutOfRangeException.ThrowIfNotEqual(uri.Collection, CollectionNsid.ListItem);
 
-
         if (!IsAuthenticated)
         {
             throw new AuthenticationRequiredException();
@@ -765,165 +571,5 @@ public partial class BlueskyAgent
             httpResponseHeaders: listEntriesResult.HttpResponseHeaders,
             atErrorDetail: new AtErrorDetail("NotFound", $"{handle} not found in list {uri}"),
             rateLimit: listEntriesResult.RateLimit);
-    }
-
-    /// <summary>
-    /// Creates a profile status for the current authenticated user to indicate a live stream.
-    /// </summary>
-    /// <param name="uri">The uri of the live stream.</param>
-    /// <param name="title">The title of the stream.</param>
-    /// <param name="description">A description of the stream.</param>
-    /// <param name="previewBlob">An optional <see cref="Blob"/> containing a preview image for the stream.</param>
-    /// <param name="durationMinutes">The optional duration of the stream in minutes.</param>
-    /// <param name="cancellationToken">A cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
-    /// <returns>The task object representing the asynchronous operation.</returns>
-    /// <exception cref="ArgumentNullException">Thrown when <paramref name="uri"/>, <paramref name="title"/>, or <paramref name="description"/> is <see langword="null"/>.</exception>
-    /// <exception cref="AuthenticationRequiredException">Thrown when the current agent is not authenticated.</exception>
-    public async Task<AtProtoHttpResult<CreateRecordResult>> CreateLiveStatus(
-        Uri uri,
-        string title,
-        string description,
-        Blob? previewBlob = null,
-        int? durationMinutes = null,
-        CancellationToken cancellationToken = default)
-    {
-        ArgumentNullException.ThrowIfNull(uri);
-        ArgumentNullException.ThrowIfNull(title);
-        ArgumentNullException.ThrowIfNull(description);
-
-        if (!IsAuthenticated)
-        {
-            throw new AuthenticationRequiredException();
-        }
-
-        Status status = new(
-            KnownStatusValues.Live,
-            embed: new EmbeddedExternal(
-                uri: uri,
-                title: title,
-                description: description,
-                thumbnail: previewBlob),
-            durationMinutes: durationMinutes,
-            createdAt: DateTimeOffset.UtcNow);
-
-        return await CreateStatus(
-            status: status,
-            cancellationToken: cancellationToken).ConfigureAwait(false);
-    }
-
-    /// <summary>
-    /// Creates a profile status for the current authenticated user.
-    /// </summary>
-    /// <param name="status">The status to set</param>
-    /// <param name="cancellationToken">A cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
-    /// <returns>The task object representing the asynchronous operation.</returns>
-    /// <exception cref="ArgumentNullException">Thrown when <paramref name="status"/> is <see langword="null"/>.</exception>
-    /// <exception cref="AuthenticationRequiredException">Thrown when the current agent is not authenticated.</exception>
-    public async Task<AtProtoHttpResult<CreateRecordResult>> CreateStatus(
-        Status status,
-        CancellationToken cancellationToken = default)
-    {
-        ArgumentNullException.ThrowIfNull(status);
-        if (!IsAuthenticated)
-        {
-            throw new AuthenticationRequiredException();
-        }
-
-        return await CreateBlueskyRecord(
-            record: status,
-            collection: CollectionNsid.Status,
-            rKey: "self",
-            validate: null,
-            cancellationToken: cancellationToken).ConfigureAwait(false);
-    }
-
-    /// <summary>
-    /// Deletes the profile status for the current authenticated user.
-    /// </summary>
-    /// <param name="cancellationToken">A cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
-    /// <returns>The task object representing the asynchronous operation.</returns>
-    /// <exception cref="AuthenticationRequiredException">Thrown when the current agent is not authenticated.</exception>
-    public async Task<AtProtoHttpResult<Commit>> DeleteStatus(
-        CancellationToken cancellationToken = default)
-    {
-        if (!IsAuthenticated)
-        {
-            throw new AuthenticationRequiredException();
-        }
-
-        AtUri statusUri = new($"at://{Did}/{CollectionNsid.Status}/self");
-        AtProtoHttpResult<Commit> deleteResult = await DeleteRecord(
-            uri: statusUri,
-            cancellationToken: cancellationToken).ConfigureAwait(false);
-
-        return deleteResult;
-    }
-
-    /// <summary>
-    /// Gets the current profile status, if any, for the authenticated user.
-    /// </summary>
-    /// <param name="cancellationToken">A cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
-    /// <returns>The task object representing the asynchronous operation.</returns>
-    /// <exception cref="AuthenticationRequiredException">Thrown when the current agent is not authenticated.</exception>
-    public async Task<AtProtoHttpResult<AtProtoRepositoryRecord<Status>>> GetStatus(CancellationToken cancellationToken = default)
-    {
-        if (!IsAuthenticated)
-        {
-            throw new AuthenticationRequiredException();
-        }
-
-        AtUri statusUri = new($"at://{Did}/{CollectionNsid.Status}/self");
-        AtProtoHttpResult<AtProtoRepositoryRecord<Status>> getResult = await GetBlueskyRecord<Status>(
-            uri: statusUri,
-            cancellationToken: cancellationToken).ConfigureAwait(false);
-
-        return getResult;
-    }
-
-    /// <summary>
-    /// Updates the status record for the authenticated user
-    /// </summary>
-    /// <param name="status">The status record to update. The record's URI authority must match the current user's decentralized identifier (DID).</param>
-    /// <param name="cancellationToken">A cancellation token that can be used to cancel the asynchronous operation.</param>
-    /// <returns>A task that represents the asynchronous update operation. The task result contains the outcome of the
-    /// update, including the updated record information.</returns>
-    /// <exception cref="AuthenticationRequiredException">Thrown if the user is not authenticated when attempting to update the status.</exception>
-    /// <exception cref="ArgumentException">Thrown if the status record's URI authority is not a valid DID or does not match the current user's DID.</exception>
-    [UnconditionalSuppressMessage(
-        "Trimming",
-        "IL2026:Members annotated with 'RequiresUnreferencedCodeAttribute' require dynamic access otherwise can break functionality when trimming application code",
-        Justification = "All types are preserved in the JsonSerializerOptions call to Get().")]
-    [UnconditionalSuppressMessage("AOT",
-        "IL3050:Calling members annotated with 'RequiresDynamicCodeAttribute' may break functionality when AOT compiling.",
-        Justification = "All types are preserved in the JsonSerializerOptions call to Get().")]
-    public async Task<AtProtoHttpResult<PutRecordResult>> UpdateStatus(
-        AtProtoRepositoryRecord<Status> status,
-        CancellationToken cancellationToken = default)
-    {
-        ArgumentNullException.ThrowIfNull(status);
-        ArgumentNullException.ThrowIfNull(status.Value);
-
-        if (!IsAuthenticated)
-        {
-            throw new AuthenticationRequiredException();
-        }
-        if (status.Uri.Authority is not Did recordDid)
-        {
-            throw new ArgumentException("Uri authority is not a DID", nameof(status));
-        }
-        if (recordDid != Did)
-        {
-            throw new ArgumentException("Uri authority does not match the current user", nameof(status));
-        }
-
-        return await PutRecord(
-            record: status.Value,
-            jsonSerializerOptions: BlueskyServer.BlueskyJsonSerializerOptions,
-            collection: CollectionNsid.Status,
-            rKey: "self",
-            validate: null,
-            swapCommit: null,
-            swapRecord: status.Cid,
-            cancellationToken: cancellationToken).ConfigureAwait(false);
     }
 }
